+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/BrowserState.kt
@@ -18,8 +18,8 @@ import mozilla.components.lib.state.State
  * The extensions here represent the default values for all [BrowserState.extensions] and can
  * be overridden per [SessionState].
  * @property media The state of all media elements and playback states for all tabs.
- * @property queuedDownloads queued downloads ([DownloadState]s) mapped to their IDs.
  * @property search the state of search for this browser state.
+ * @property downloads Downloads ([DownloadState]s) mapped to their IDs.
  */
 data class BrowserState(
     val tabs: List<TabSessionState> = emptyList(),
@@ -28,6 +28,6 @@ data class BrowserState(
     val containers: Map<String, ContainerState> = emptyMap(),
     val extensions: Map<String, WebExtensionState> = emptyMap(),
     val media: MediaState = MediaState(),
-    val queuedDownloads: Map<Long, DownloadState> = emptyMap(),
+    val downloads: Map<Long, DownloadState> = emptyMap(),
     val search: SearchState = SearchState()
 ) : State